+++ R/snapshot-file.R
@@ -25,6 +25,10 @@
 #'   `new` arguments. By default this is `compare_file_binary`. Set it
 #'   to `compare_file_text` to compare files line-by-line, ignoring
 #'   the difference between Windows and Mac/Linux line endings.
+#' @param variant If not-`NULL`, results will be saved in
+#'   `_snaps/{variant}/{test}/{name}.{ext}`. This allows you to create
+#'   different snapshots for different scenarios, like different operating
+#'   systems or different R versions.
 #' @inheritParams expect_snapshot
 #'
 #' @section Announcing snapshots:
@@ -80,11 +84,14 @@ expect_snapshot_file <- function(path,
                                  name = basename(path),
                                  binary = lifecycle::deprecated(),
                                  cran = FALSE,
-                                 compare =  compare_file_binary) {
+                                 compare =  compare_file_binary,
+                                 variant = NULL) {
   edition_require(3, "expect_snapshot_file()")
   if (!cran && !interactive() && on_cran()) {
     skip("On CRAN")
   }
+
+  check_variant(variant)
 
   snapshotter <- get_snapshotter()
   if (is.null(snapshotter)) {
@@ -102,7 +109,7 @@ expect_snapshot_file <- function(path,
   }
 
   lab <- quo_label(enquo(path))
-  equal <- snapshotter$take_file_snapshot(name, path, compare)
+  equal <- snapshotter$take_file_snapshot(name, path, compare, variant = variant)
   hint <- snapshot_hint(snapshotter$file, name)
 
   expect(
@@ -157,31 +164,6 @@ snapshot_file_equal <- function(snap_test_dir, snap_name, path, file_equal = com
     testthat_warn(paste0("Adding new file snapshot: '", cur_path, "'"))
     TRUE
   }
-}
-
-snapshot_file_outdated <- function(snap_dir, tests_seen = character(), snaps_seen = character()) {
-  # 1) Entire test file deleted/moved. Need to delete _snaps/test/
-  # Recognise because missing from tests_seen
-  tests <- list.dirs(snap_dir, recursive = FALSE)
-  tests_outdated <- tests[!basename(tests) %in% tests_seen]
-
-  # 2) Single test deleted. Need to delete _snaps/test/foo.txt
-  # Recognise because missing from snap_seens
-  snaps <- dir(snap_dir, recursive = TRUE, full.names = TRUE)
-  snap_names <- dir(snap_dir, recursive = TRUE, full.names = FALSE)
-
-  # Remove all files in the root snapshot directory - those are managed by
-  # snapshot_outdated()
-  base <- snaps %in% dir(snap_dir, full.names = TRUE)
-  snaps <- snaps[!base]
-  snap_names <- snap_names[!base]
-
-  snaps_outdated <- snaps[
-    !snap_names %in% c(snaps_seen, new_name(snaps_seen)) &
-    dirname(snap_names) %in% tests_seen
-  ]
-
-  unique(c(tests_outdated, snaps_outdated))
 }
 
 # Helpers -----------------------------------------------------------------