+++ tests/testthat/test-snapshot-manage.R
@@ -1,5 +1,5 @@
 test_that("informs about files being accepted", {
-  path <- local_snapshot_dir(c("a", "b"))
+  path <- local_snapshot_dir(c("a", "a.new", "b", "b.new"))
 
   expect_snapshot(snapshot_accept(path = path))
   expect_equal(dir(file.path(path, "_snaps")), c("a.md", "b.md"))
@@ -8,7 +8,7 @@ test_that("informs about files being accepted", {
 })
 
 test_that("can accept specific files", {
-  path <- local_snapshot_dir(c("a", "b"))
+  path <- local_snapshot_dir(c("a", "a.new", "b", "b.new"))
   expect_snapshot(snapshot_accept("a", path = path))
   expect_equal(dir(file.path(path, "_snaps")), c("a.md", "b.md", "b.new.md"))
 })